Feature: Include enabled security response headers
It is assumed that scans performed before this commit do not have any security response header enabled, to avoid problems with scans performed before this date.

It is not the best option but I think it is one of the most polite :). Any analysis performed after this commit will save in the analysis history the total of enabled security response headers.