Feature: Bump required python version to 3.9
https://github.com/rfc-st/humble/issues/25

Unfortunately vermin, the tool I used to check the minimum Python version required to use 'humble' is not updated with rules for (at least) Python 3.12 ... and in fact it may not be updated for earlier versions of Python either. So its results, at least for those versions of Python, are not reliable.

So, apart from stopping using vermin, I will indicate from now on that 'humble' is compatible with (at least) the minimum supported version of Python, which is currently 3.9: https://devguide.python.org/versions/

In the short/medium term I will investigate unit tests on github for that version of Python, trying to get ahead of any bugs that may arise.